requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
SQLAlchemy>=1.4.23
psycopg2-binary>=2.9.3
python-dotenv>=0.19.0
bs4>=0.0.1
requests>=2.26.0
pydantic>=1.8.2
python-dateutil>=2.8.2
colorama>=0.4.4
supabase>=1.0.3